server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
opencv-python==4.10.0.84
numpy==1.26.4
mediapipe==0.10.14
python-dotenv==1.0.1
pydantic==2.7.4
pytest==8.2.2
